+++ shadowsocks-csharp/Encryption/SodiumEncryptor.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Shadowsocks.Encrypt
+namespace Shadowsocks.Encryption
 {
     public class SodiumEncryptor
         : IVEncryptor, IDisposable
